* Fixed bug introduced in r2752 (PRACK support broken).
git-svn-id: svn://svn.minisip.org/minisip/trunk@2754 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/sip/SipDialogVoipClient.cxx
@@ -790,7 +790,7 @@ bool SipDialogVoipClient::handleRel1xx( MRef<SipResponse*> resp ){
 		
 	// First reliable provisional response
 	// Next in-order reliable provisional response
-	if( !(dialogState.rseqNo == (uint32_t)-1 ||
+	if( !(dialogState.rseqNo == -1 ||
 	      dialogState.rseqNo > rseqNo ) )
 		return false;
